Use temporal.download for downloading test server
