+++ packages/testing/scripts/download-test-server.mjs
@@ -2,10 +2,29 @@ import stream from 'node:stream';
 import util from 'node:util';
 import zlib from 'node:zlib';
 import fs from 'node:fs';
+import os from 'node:os';
 import got from 'got';
 import tar from 'tar-stream';
 import unzipper from 'unzipper';
-import { outputPath, systemArch, systemPlatform } from './common.mjs';
+import { URL, fileURLToPath } from 'node:url';
+import pkg from '../package.json' assert { type: 'json' };
+
+const platformMapping = { darwin: 'darwin', linux: 'linux', win32: 'windows' };
+const archAlias = { x64: 'amd64', arm64: 'aarch64' };
+
+const systemPlatform = platformMapping[os.platform()];
+if (!systemPlatform) {
+  throw new Error(`Unsupported platform ${os.platform()}`);
+}
+
+const systemArch = archAlias[os.arch()];
+if (!systemArch) {
+  throw new Error(`Unsupported architecture ${os.arch()}`);
+}
+
+const ext = systemPlatform === 'windows' ? '.exe' : '';
+const assetExt = systemPlatform === 'windows' ? 'zip' : 'tar.gz';
+const outputPath = fileURLToPath(new URL(`../test-server${ext}`, import.meta.url));
 
 try {
   if (fs.statSync(outputPath).isFile) {
@@ -20,64 +39,28 @@ try {
 
 const pipeline = util.promisify(stream.pipeline);
 
-const defaultHeaders = {
-  'User-Agent': '@temporalio/testing installer',
+const options = {
+  headers: {
+    'User-Agent': '@temporalio/testing installer',
+  },
+  searchParams: { 'sdk-name': 'typescript', 'sdk-version': pkg.version },
 };
 
-const { GITHUB_TOKEN } = process.env;
+// TODO: use "real" arch when we get arm builds for test server
+const assetUrl = `https://temporal.download/temporal-test-server/temporal-test-server_default_${systemPlatform}_amd64.${assetExt}`;
+console.log('Downloading test server', { assetUrl, outputPath });
 
-if (GITHUB_TOKEN) {
-  console.log(`Using GITHUB_TOKEN`);
-  defaultHeaders['Authorization'] = `Bearer ${GITHUB_TOKEN}`;
-}
-
-const latestReleaseRes = await got('https://api.github.com/repos/temporalio/sdk-java/releases/latest', {
-  headers: {
-    ...defaultHeaders,
-    Accept: 'application/vnd.github.v3+json',
-  },
-}).json();
-
-function findTestServerAsset(assets) {
-  for (const asset of assets) {
-    const m = asset.name.match(/^temporal-test-server_[^_]+_([^_]+)_([^.]+)\.(?:zip|tar.gz)$/);
-    if (m) {
-      const [_, assetPlatform, _assetArch] = m;
-      if (assetPlatform === systemPlatform) {
-        // TODO: assetArch === systemArch (no arm builds for test server yet)
-        return asset;
-      }
-    }
-  }
-  throw new Error(`No prebuilt test server for ${systemPlatform}-${systemArch}`);
-}
-
-const asset = findTestServerAsset(latestReleaseRes.assets);
-console.log('Downloading test server', { asset: asset.name, outputPath });
-
-if (asset.content_type === 'application/x-gzip' || asset.content_type === 'application/x-gtar') {
+if (assetExt === 'tar.gz') {
   const extract = tar.extract();
   extract.on('entry', (_headers, stream, next) => {
     stream.pipe(fs.createWriteStream(outputPath));
     next();
   });
-  await pipeline(
-    got.stream(asset.browser_download_url, {
-      headers: {
-        ...defaultHeaders,
-      },
-    }),
-    zlib.createGunzip(),
-    extract
-  );
+  await pipeline(got.stream(assetUrl, options), zlib.createGunzip(), extract);
   await fs.promises.chmod(outputPath, 0o755);
-} else if (asset.content_type === 'application/zip') {
+} else if (assetExt === 'zip') {
   got
-    .stream(asset.browser_download_url, {
-      headers: {
-        ...defaultHeaders,
-      },
-    })
+    .stream(assetUrl, options)
     .pipe(unzipper.Parse())
     .on('entry', (entry) => {
       if (entry.type === 'File') {
@@ -86,6 +69,4 @@ if (asset.content_type === 'application/x-gzip' || asset.content_type === 'appli
         entry.autodrain();
       }
     });
-} else {
-  throw new Error(`Unexpected content type for Test server download: ${asset.content_type}`);
 }
